moved main script into main function
